pyproject.toml
[project]
name = "testfrontend"
version = "0.1.0"
description = ""
authors = [
    {name = "Zwick Vitaly",email = "zwickvitaly@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all] (>=0.115.8,<0.116.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
